[feat] #27 온보딩화면과 목표변경화면의 공통 컴포넌트 추출, 패키지 정리
